requirements.txt
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
decorator==4.4.2
future==1.0.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
huepy==1.2.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
instabot==0.117.0
instagrapi==2.1.1
mock==5.1.0
moviepy==1.0.3
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.25.0
rsa==4.9
schedule==1.2.1
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0
annotated-types==0.6.0
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
Flask==3.0.3
future==1.0.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
huepy==1.2.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
instabot==0.117.0
instagrapi==2.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mock==5.1.0
moviepy==1.0.3
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.25.0
rsa==4.9
schedule==1.2.1
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
